cleanup logic for logs
