+++ config.json
@@ -2,7 +2,7 @@
     "logging": {
         "level": "INFO",
         "name": "r2r",
-        "database": "demo_logs_v1"
+        "database": "demo_logs_v2"
     },
     "embedding": {
         "model": "text-embedding-3-small",
